requirements.txt
python-dotenv==1.0.1
packaging~=25.0
fastapi~=0.115.14
PyYAML~=6.0.2
openai~=1.93.0
slack_sdk~=3.35.0
google~=3.0.0
protobuf~=5.29.5
seaborn~=0.13.2
matplotlib~=3.10.3
pandas~=2.3.0
router~=0.1
RapidFuzz~=3.13.0
uvicorn~=0.35.0
google-cloud-bigquery==3.17.2